requirements.txt
django==3.1.7
djangorestframework==3.12.2
dj-stripe==2.4.3
